+++ paper/paper.bib
@@ -8,6 +8,7 @@
   publisher="Springer Singapore",
   address="Singapore",
   pages="65--74",
+  doi="10.1007/978-981-13-0860-4_6",
   url="https://doi.org/10.1007/978-981-13-0860-4_6",
 }
 
@@ -51,6 +52,7 @@
   pages={459--471},
   year={2007},
   publisher={Springer},
+  doi={10.1007/s10898-007-9149-x},
   url={https://doi.org/10.1007/s10898-007-9149-x},
 }
 
@@ -64,6 +66,7 @@
   pages={284--302},
   year={2008},
   publisher={IEEE},
+  doi={10.1109/tevc.2008.925798},
   url={https://doi.org/10.1109/TEVC.2008.925798}
 }
 
@@ -77,6 +80,7 @@
   pages={407--419},
   year={2017},
   publisher={Elsevier},
+  doi={10.1016/j.asoc.2017.01.008},
   url={https://doi.org/10.1016/j.asoc.2017.01.008}
 }
 
@@ -113,6 +117,7 @@
   pages={182--197},
   year={2002},
   publisher={IEEE},
+  doi={10.1109/4235.996017},
   url={https://doi.org/10.1109/4235.996017},
 }
 
@@ -167,7 +172,8 @@
   number  = {1},
   pages   = {43--58},
   year    = {2020},
-  DOI     = {https://doi.org/10.17093/alphanumeric.576919}
+  DOI     = {10.17093/alphanumeric.576919},
+  url     = {https://doi.org/10.17093/alphanumeric.576919}
 }
 
 @article{Satman2021JMcDM,
