+++ test/multi-objective.jl
@@ -32,6 +32,7 @@ end
                 SPEA2(options=options),
                 SMS_EMOA(N = 50, n_samples=500, options=options),
                 NSGA2(options=options),
+                NSGA2(N=59,options=options),
                 MOEAD_DE(gen_ref_dirs(nobjectives, npartitions), options=Options( seed = 1, iterations = 500)),
                 NSGA3(options=options),
               ]
