### 2022-08-08 9aea4f4
Create FUNDING.yml
