+++ .github/FUNDING.yml
@@ -1,0 +1,4 @@
+# These are supported funding model platforms
+
+github: jmejia8
+ko_fi: jmejia
