+++ v3.2.10/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"references/#References","page":"References","title":"References","text":"","category":"section"},{"location":"references/","page":"References","title":"References","text":"","category":"page"},{"location":"tutorials/simple-tutorial/#Getting-Started","page":"Getting Started","title":"Getting Started","text":"","category":"section"},{"location":"tutorials/simple-tutorial/","page":"Getting Started","title":"Getting Started","text":"After reading this tutorial you'll become an expert in using the Metaheuristics module.","category":"page"},{"location":"tutorials/simple-tutorial/#Minimization-Problem","page":"Getting Started","title":"Minimization Problem","text":"","category":"section"},{"location":"tutorials/simple-tutorial/","page":"Getting Started","title":"Getting Started","text":"Assume you want to optimize the following minimization problem:","category":"page"},{"location":"tutorials/simple-tutorial/","page":"Getting Started","title":"Getting Started","text":"Minimize:","category":"page"},{"location":"tutorials/simple-tutorial/","page":"Getting Started","title":"Getting Started","text":"f(x) = 10D + sum_i=1^D x_i^2 - 10cos(2pi x_i)","category":"page"},{"location":"tutorials/simple-tutorial/","page":"Getting Started","title":"Getting Started","text":"where xin -5 5^D, that is, each coordinate in x is between -5 and 5. Use D=10.","category":"page"},{"location":"tutorials/simple-tutorial/","page":"Getting Started","title":"Getting Started","text":"Note that the global optimum is obtained when x_i = 0 for all i. Thus, min f(x) = 0.","category":"page"},{"location":"tutorials/simple-tutorial/","page":"Getting Started","title":"Getting Started","text":"Objective function:","category":"page"},{"location":"tutorials/simple-tutorial/","page":"Getting Started","title":"Getting Started","text":"f(x) = 10length(x) + sum( x.^2 - 10cos.(2pi*x) )","category":"page"},{"location":"tutorials/simple-tutorial/","page":"Getting Started","title":"Getting Started","text":"Bounds:","category":"page"},{"location":"tutorials/simple-tutorial/","page":"Getting Started","title":"Getting Started","text":"bounds = [-5ones(10) 5ones(10)]'","category":"page"},{"location":"tutorials/simple-tutorial/#Providing-Information","page":"Getting Started","title":"Providing Information","text":"","category":"section"},{"location":"tutorials/simple-tutorial/","page":"Getting Started","title":"Getting Started","text":"Since the optimum is known, then we can provide this information to the optimizer.","category":"page"},{"location":"tutorials/simple-tutorial/","page":"Getting Started","title":"Getting Started","text":"information = Information(f_optimum = 0.0)","category":"page"},{"location":"tutorials/simple-tutorial/#Common-Settings","page":"Getting Started","title":"Common Settings","text":"","category":"section"},{"location":"tutorials/simple-tutorial/","page":"Getting Started","title":"Getting Started","text":"Usually, users could require to limit the number of generations/iterations or the number of function evaluations. To do that, let's assume that the metaheuristic should evaluate at most 9000D times the objective function. Moreover, since information is provided, then we can set the desired accuracy (f(x) - f(x^*) ) to 10^-5.","category":"page"},{"location":"tutorials/simple-tutorial/","page":"Getting Started","title":"Getting Started","text":"options = Options(f_calls_limit = 9000*10, f_tol = 1e-5)","category":"page"},{"location":"tutorials/simple-tutorial/#Choose-a-Metaheuristic","page":"Getting Started","title":"Choose a Metaheuristic","text":"","category":"section"},{"location":"tutorials/simple-tutorial/","page":"Getting Started","title":"Getting Started","text":"Metaheuristics.jl provides different metaheuristics 
[…]
