+++ versions.js
@@ -6,5 +6,5 @@ var DOC_VERSIONS = [
   "v2.2",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v3.2.9";
+var DOCUMENTER_NEWEST = "v3.2.10";
 var DOCUMENTER_STABLE = "stable";
