### 2022-08-22 52c2fee
add an explicit example for MCCGA
