### 2022-09-01 #4723 @editorialbot
[REVIEW]: Metaheuristics: A Julia Package for Single- and Multi-Objective Optimization
**Submitting author:** @jmejia8 (<a href="http://orcid.org/0000-0002-0367-2967">Jesus-Adolfo Mejia-de-Dios</a>)
**Repository:** https://github.com/jmejia8/Metaheuristics.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v3.2.12
