### 2022-09-01 #4723 @idoby
@jmejia8, first of all, thank you for submitting your work 📦 .

🎉 I want first to applaud the substantial effort that coding, testing and documenting this package must have required. From a first look, the package seems to be well-coded, testing coverage seems to be within the sweet-spot range and the documentation seems well thought-out, is aesthetically pleasing and organized well.

👌  Minor improvements should be made, in my opinion, to some of the software materials as well as the paper itself. I will open issues in the original repository when time permits.

🎓  In the meanwhile, please add a section to your paper explaining how your package relates to other Julia packages relating to constrained optimization, evolutionary algorithms, heuristic search, etc. Many relevant packages are listed here: https://www.juliaopt.org/packages/. I also encourage you to discuss packages from other popular languages, e.g. Python and Matlab, although this is less important, in my opinion.

Keep up the good work!
