+++ src/Metaheuristics.jl
@@ -24,6 +24,7 @@ export UniformCrossover, OrderCrossover, SBX
 export BitFlipMutation, SlightMutation, PolynomialMutation
 export GenerationalReplacement, ElitistReplacement
 export sample, LatinHypercubeSampling, Grid 
+export εDE
 
 include("externals.jl")
 
