+++ docs/references.bib
@@ -246,3 +246,14 @@ url="https://doi.org/10.1007/978-1-4615-3612-3_4"
   pages={102-116},
 
   doi={10.1109/TEVC.2020.3004012}}
+
+
+@INPROCEEDINGS{Takahama2006Constrained,
+  author={Takahama, T. and Sakai, S.},
+  booktitle={2006 IEEE International Conference on Evolutionary Computation}, 
+  title={Constrained Optimization by the ε Constrained Differential Evolution with Gradient-Based Mutation and Feasible Elites}, 
+  year={2006},
+  volume={},
+  number={},
+  pages={1-8},
+  doi={10.1109/CEC.2006.1688283}}
