### 2022-09-05 be31c4b
add docstring on epsilonDE
