### 2022-09-23 693ad3f
Proposed changes to paper.md
Rewrote some of the paper file to be clearer and more accurate (in my humble opinion, of course).
