+++ paper/paper.md
@@ -25,27 +25,25 @@ bibliography: paper.bib
 
 # Summary
 
-`Metaheuristics` is a `Julia` package that implements metaheuristic algorithms for solving global optimization problems that can contain constraints, and either single or multiple objectives.
-The package implements an easy-to-use API that let testing without extra configurations among the implemented optimizers, i.e., to `optimize` an objective function $f(x)$ with $x$ in the corresponding `bounds` using optimizer `ECA`, the user can write `optimize(f, bounds, ECA())` to approximate an optimal solution.
-Moreover, `Metaheuristics` provides common features required by the Evolutionary Computing community such as challenging test problems, performance indicators, and other interest utility functions devoted to performance improvement.
-This paper presents the main features and some examples to illustrate the usage of this package for optimization.
+Metaheuristics is a Julia package that implements metaheuristic algorithms for solving global optimization problems that can contain constraints, and either single or multiple objectives.
+The package exposes an easy-to-use API to enable testing without requiring length configuration for choosing among the implemented optimizers. For example, to optimize an objective function $f(x)$ where solution $x$ is within the corresponding bounds using the Evolutionary Centers Algorithm optimizer, the user can call `optimize(f, bounds, ECA())`.
+Moreover, Metaheuristics provides common features required by the evolutionary computing community such as challenging test problems, performance indicators, and other notable utility functions.
+This paper presents the main features followed by some examples to illustrate the usage of this package for optimization.
 
 
 # Statement of need
 
-
-
-Real-world problems require sophisticated methodologies to provide feasible and efficient solutions to hard problems.
-Metaheuristics are algorithms proposed to approximate optimal solutions in a short time to optimization problems with unknown mathematical properties. The package implements state-of-the-art algorithms for constrained, multi-, and many-objective optimization.
-Besides, many other utility functions have been implemented in the package, e.g., performance indicators for performance assessment, scalable benchmark problems, constraint handling techniques, and multi-criteria decision-making methodologies.
-To the best knowledge of the authors, `Metaheuristics` is the first package in `Julia` containing those ready-to-use features with a uniform API. The package can be used for both academic and industrial purposes due to it is distributed through a flexible license.
-
-
+Real-world problems often require sophisticated methods to solve.
+Metaheuristics are algorithms that approximate optimal solutions quickly where not all of the mathematical properties of the problem are known.
+Our package implements state-of-the-art algorithms for constrained, multi-, and many-objective optimization.
+It also includes many other utility functions such as performance indicators, scalable benchmark problems, constraint handling techniques, and multi-criteria decision-making methodologies.
+To the best knowledge of the authors, Metaheuristics is the first package in Julia containing those ready-to-use metahuristic algorithms with a uniform API.
+The package can be used for both academic and industrial purposes due to its flexible license.
 
 # Main Features
 
-This part aims to describe the most important features included in the package.
[…]
