### 2022-09-28 jmejia8/Metaheuristics.jl#36 @PaulDebus
JOSS Review Tracker
Hey,

I am currently working on my [review for JOSS](https://github.com/openjournals/joss-reviews/issues/4723) on the paper you submitted. In this issue, I will collect all the things I would like to discuss with you, so we can keep all in one place.

In general, I think the work is very well-thought-out and only minor things remain. You obviously put a lot of effort in and produced a high quality project.

- [x] Consider an example with more reproducible results. I ran the example you show both in your paper and the repo and the optimization does not always find the `0` solution (I tried with 1000 optimization runs and only in ~50% of them I found the true minimum). To make it more intuitive for new people, maybe you can find an example that more reliably produces the optimal result? Alternatively, you could add a small note with an explanation.
- [x] Discuss other Julia packages for optimization and how your work is different and how it integrates with them and the larger ecosystem.
- [x] Please run a quick spell check, some minor errors and unusual formulations still exist.

Once these topics here are resolved, I can finish my review and recommend publication.

Cheers,
Paul
