### 2022-09-28 #4723 @PaulDebus
Hey @jmejia8

thanks for contributing this interesting work, I am having a good time working on my review and even learned some things ;)

It clearly shows you put a lot of effort in to produce a high quality software tool, that has some clear applications and can be useful. The documentation is very good and shows all the features and how to use them - really nice.

I have opened a [separate issue](https://github.com/jmejia8/Metaheuristics.jl/issues/36) in your repo for the small aspects I think should be addressed before publication. Overall, I agree with @idoby - especially with regard to other optimization packages and the Julia ecosystem. Please add a dedicated section on how your work relates to other tools.

Best
Paul
