### 2022-09-28 #4723 @idoby
@vissarion the bot doesn't seem to take commands if there are multiple lines in the comment, is this intentional? Would just make for longer/noisier threads
