+++ paper/paper.bib
@@ -190,14 +190,24 @@
 }
 
 
+
 @Inbook{sebastian2022efficient,
-  author="de-la-Cruz-Martínez, S.J and Mejía-de-Dios, J.A. and  Mezura-Montes E.",
-  title="Efficient Archiving Method for Handling Preferences in Constrained Multiobjective Evolutionary Optimization",
-  bookTitle="Handbook on Decision Making: Trends and Challenges in Intelligent Decision Support Systems",
-  year="2022",
-  publisher="Springer-Verlag",
-  pages="1--30",
-  volume="3",
+author="de-la-Cruz-Mart{\'i}nez, Sebasti{\'a}n Jos{\'e}
+and Mej{\'i}a-de-Dios, Jes{\'u}s Adolfo
+and Mezura-Montes, Efr{\'e}n",
+editor="Zapata-Cortes, Julian Andres
+and S{\'a}nchez-Ram{\'i}rez, Cuauht{\'e}moc
+and Alor-Hern{\'a}ndez, Giner
+and Garc{\'i}a-Alcaraz, Jorge Luis",
+title="Efficient Archiving Method for Handling Preferences in Constrained Multi-objective Evolutionary Optimization",
+bookTitle="Handbook on Decision Making: Volume 3: Trends and Challenges in Intelligent Decision Support Systems",
+year="2023",
+publisher="Springer International Publishing",
+address="Cham",
+pages="93--119",
+isbn="978-3-031-08246-7",
+doi="10.1007/978-3-031-08246-7_5",
+url="https://doi.org/10.1007/978-3-031-08246-7_5"
 }
 
 
@@ -262,3 +272,116 @@ url="https://doi.org/10.1007/978-1-4615-3612-3_4"
   publisher={Springer}
 }
 
+
+@article{mogensen2018optim,
+  author  = {Mogensen, Patrick Kofod and Riseth, Asbj{\o}rn Nilsen},
+  title   = {Optim: A mathematical optimization package for {Julia}},
+  journal = {Journal of Open Source Software},
+  year    = {2018},
+  volume  = {3},
+  number  = {24},
+  pages   = {615},
+  doi     = {10.21105/joss.00615}
+}
+
+@software{RobertFeldtBBO,
+  author = {Robert Feldt},
+  title = {{BlackBoxOptim.jl: Black-box optimization for Julia}},
+  year = {2022},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/robertfeldt/BlackBoxOptim.jl}
+}
+
+
+@software{artewilde2021,
+  author       = {Art Wilde and
+                  Daniel Thiem and
+                  Leo and
+                  Abhinav Gupta and
+                  Andrew Haselgrove and
+                  Daniel Molina and
+                  Honeypot95 and
+                  matago and
+                  Rogerluo and
+                  Valentin Churavy},
+  title        = {Evolutionary.jl},
+  month        = jul,
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {v0.10.0},
+  doi          = {10.5281/zenodo.5110647},
+  url          = {https://doi.org/10.5281/zenodo.5110647}
+}
+
+
+@software{CMAEvolutionStrategy,
+  author = {Johanni Brea},
+  title = {{CMAEvolutionStrategy.jl: CMA Evolution Strategy in Julia}},
+  year = {2022},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/jbrea/CMAEvolutionStrategy.jl}
+}
+
+
+@ARTICLE{pymoo,
+    author={J. {Blank} and K. {Deb}},
+    journal={IEEE Access},
+    title={pymoo: Multi-Objective Optimization in Python},
+    year={2020},
[…]
