+++ paper/paper.bib
@@ -326,13 +326,15 @@ url="https://doi.org/10.1007/978-1-4615-3612-3_4"
 
 
 @ARTICLE{pymoo,
-    author={J. {Blank} and K. {Deb}},
-    journal={IEEE Access},
-    title={pymoo: Multi-Objective Optimization in Python},
-    year={2020},
-    volume={8},
-    number={},
-    pages={89497-89509},
+  author={Blank, Julian and Deb, Kalyanmoy},
+  journal={IEEE Access}, 
+  title={Pymoo: Multi-Objective Optimization in Python}, 
+  year={2020},
+  volume={8},
+  number={},
+  pages={89497-89509},
+  doi={10.1109/ACCESS.2020.2990567},
+  url={https://doi.org/10.1109/ACCESS.2020.2990567}
 }
 
 @article{PlatEMO,
