### 2022-10-04 #4723 @vissarion
Hi all, please allow me to make a brief status update here.

@PaulDebus is waiting for @jmejia8 to address the issue opened in the repository https://github.com/jmejia8/Metaheuristics.jl/issues/36

@idoby you have some unchecked boxes in your review checklist. Are you waiting for some issue to be addressed by the author?

@JustinSGray any updates on your review progress?
