### 2022-10-04 #4723 @idoby
@vissarion I have updated my checklist according to the latest revision of the paper from yesterday. I think the review has already contributed to the improvement of the paper and package and as far as I'm concerned, all issues have been addressed.

@jmejia8 I have just one last comment - I think that printing out status information out of your library is a usability obstacle for a lot of potential users. You should consider avoiding prints within your library since I might be using a custom logger with my own custom format, sending events to the cloud not even monitoring stdout, etc. Or it might simply get in the way of the output I want to see. I suggest returning important information in a structured way so that they are also inspectable with a debugger, and then providing a mechanism to turn these objects into a pretty string, if the user wants to.

However, I don't think this issue should block acceptance of Metaheuristics into JOSS. I will leave this decision up to the authors.

Great work, thank you everyone!
