### 2022-10-04 jmejia8/Metaheuristics.jl#36 @jmejia8
@PaulDebus, thank you very much for the important comments. I addressed your suggestions in https://github.com/jmejia8/Metaheuristics.jl/commit/a6bfaff3b154ab6a45d022ec0d09cf86bb1fbf1a

- A minor update in the example has been considered to guarantee reproducibility.
- The literature review and discussion of related software (in Julia and other programming languages) have been added.
- A spell check has been performed. Please note that @idoby updated the document via https://github.com/jmejia8/Metaheuristics.jl/pull/35

If you have any other comments in this regard, please let me know.
