### 2022-10-04 #4723 @jmejia8
First of all, thank you all for taking the time to write this review.

@PaulDebus, I just commented on the issue https://github.com/jmejia8/Metaheuristics.jl/issues/36

@idoby, your last comment is very important. As far as I understand, Julia can handle the [I/O](https://docs.julialang.org/en/v1/base/io-network/) depending on the context. The package overloads the [show](https://docs.julialang.org/en/v1/base/io-network/#Base.show-Tuple{IO,%20Any,%20Any}) method to [display](https://docs.julialang.org/en/v1/base/io-network/#Base.Multimedia.display) the status regarding the context. Actually, the [optimize](https://jmejia8.github.io/Metaheuristics.jl/stable/api/) method has a parameter named `logger`, which receives the status at the end of each iteration. We can discuss this in a separate issue.
