### 2022-10-04 #4723 @idoby
@jmejia8 Up to you. Ultimately, it's a matter of preference. I always prefer to decouple I/O from processing and use structured information instead of unstructured information.
It's much easier to print out structured information than to have to parse in I/O after the fact, if/when the need arises (even if Julia can handle different MIME types).
