### 2022-10-04 88f4d27
Update paper.md
Small typo and incomplete sentence fix
