### 2022-10-04 jmejia8/Metaheuristics.jl#36 @PaulDebus
Hey @jmejia8

Thanks for addressing my suggestions. This satisfies the questions I had, and I am happy to finish my review. I only opened one tiny pull request to fix two things in the paper:

https://github.com/jmejia8/Metaheuristics.jl/compare/paper...PaulDebus:Metaheuristics.jl:patch-1

Cool contribution, good job!
