### 2022-10-04 #4723 @PaulDebus
Hey,

thanks for addressing my questions and concerns over at https://github.com/jmejia8/Metaheuristics.jl/issues/36
This resolves the remaining points for me, and I updated my checklist accordingly.

For me, this paper can be accepted (possibly after my last tiny pull request to fix a typo is merged :wink: ) to JOSS

Good work to the authors and reviewers, especially @idoby for your good input and engagement and @jmejia8 for the good contribution!
