+++ test/box-constrained.jl
@@ -43,6 +43,7 @@ import Optim
                    SA(options = options, information = information),
                    WOA(options = options, information = information),
                    MCCGA(options = options_3, information = information),
+                   Restart(WOA(;options, information)),
                   ]
 
         # CGSA
