### 2022-10-20 c6d3fe6
update compat for UnicodePlots
