+++ Project.toml
@@ -15,7 +15,7 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 UnicodePlots = "b8865327-cd53-5732-bb35-84acbb429228"
 
 [compat]
-UnicodePlots = "^2.6.0"
+Distances = "^0.10.7"
 Requires = "^1.3.0"
-Distances="^0.10.7"
+UnicodePlots = "2.6, 3"
 julia = "^1.5"
