+++ versions.js
@@ -6,5 +6,5 @@ var DOC_VERSIONS = [
   "v2.2",
   "dev",
 ];
-var DOCUMENTER_NEWEST = "v3.2.11";
+var DOCUMENTER_NEWEST = "v3.2.12";
 var DOCUMENTER_STABLE = "stable";
