### 2022-10-22 #4723 @jmejia8
Thank you all for the time you spent on our submission. @vissarion, below is the requested information:

- Tagged release: [v3.2.12](https://github.com/jmejia8/Metaheuristics.jl/releases/tag/v3.2.12)
- Zenodo DOI: [10.5281/zenodo.7239440](https://doi.org/10.5281/zenodo.7239440)
