### 2022-10-23 4779ea7
(#43) Replacing Nelder-Mead with Hooke-Jeeves in MCCGA
