### 2022-10-23 1d5ee88
remove Optim dependency from tests
