+++ test/multi-objective.jl
@@ -44,7 +44,7 @@ end
             show(IOBuffer(), "text/plain", result.population)
             show(IOBuffer(), "text/html", result.population)
             show(IOBuffer(), result.population[1])
-            @test Metaheuristics.PerformanceIndicators.igd(result.population, pf) <= 0.2
+            @test Metaheuristics.PerformanceIndicators.igd(result.population, pf) <= 0.5
             @test Metaheuristics.PerformanceIndicators.spacing(result) < 0.5
             @test Metaheuristics.PerformanceIndicators.covering(pf, result.population) <= 1.0
             @test Metaheuristics.PerformanceIndicators.covering(result, result) ≈ 0.0
