### 2022-10-25 #4723 @jmejia8
@vissarion, the archive was automatically generated after tagging the release. The listed co-authors have been contributors who have made minor changes to the package. The metadata in the Zenodo archive has been updated to match the paper data.
