### 2022-10-26 97b57d3
minor changes in bib
