+++ paper/paper.bib
@@ -3,7 +3,7 @@
   author={Mejía-de-Dios, J.-A. and Mezura-Montes, E.},
   editor="Deep, Kusum and Jain, Madhu and Salhi, Said",
   title="A New Evolutionary Optimization Method Based on Center of Mass",
-  bookTitle="Decision Science in Action: Theory and Applications of Modern Decision Analytic Optimisation",
+  bookTitle="Decision {S}cience in {A}ction: Theory and {A}pplications of {M}odern {D}ecision {A}nalytic {O}ptimisation",
   year="2019",
   publisher="Springer Singapore",
   address="Singapore",
@@ -20,7 +20,7 @@
   and Sná Sel, Václav
   and Abraham, Ajith",
   title="Differential Evolution",
-  bookTitle="Handbook of Optimization: From Classical to Modern Approach",
+  bookTitle="Handbook of {O}ptimization: From {C}lassical to {M}odern {A}pproach",
   year="2013",
   publisher="Springer Berlin Heidelberg",
   address="Berlin, Heidelberg",
@@ -58,7 +58,7 @@
 
 
 @article{LiZhang2008,
-  title={Multiobjective optimization problems with complicated Pareto sets, MOEA/D and NSGA-II},
+  title={Multiobjective optimization problems with complicated {P}areto sets, MOEA/D and NSGA-II},
   author={Li, H. and Zhang, Q.},
   journal={IEEE transactions on evolutionary computation},
   volume={13},
@@ -88,7 +88,7 @@
 @incollection{Van1987,
   title={Simulated annealing},
   author={Van L., Peter J.M. and Aarts, E.H.L.},
-  booktitle={Simulated annealing: Theory and applications},
+  booktitle={Simulated {A}nnealing: {T}heory and {A}minor cpplications},
   pages={7--15},
   year={1987},
   publisher={Springer},
@@ -128,7 +128,7 @@
 
   journal={IEEE Transactions on Evolutionary Computation}, 
 
-  title={An Evolutionary Many-Objective Optimization Algorithm Using Reference-Point-Based Nondominated Sorting Approach, Part I: Solving Problems With Box Constraints}, 
+  title={An Evolutionary Many-Objective Optimization Algorithm Using Reference-Point-Based Nondominated Sorting Approach, Part {I}: Solving Problems With Box Constraints}, 
 
   year={2014},
 
@@ -150,12 +150,12 @@
   pages = {62--76},
   author = {Michael Emmerich and Nicola Beume and Boris Naujoks},
   title = {An {EMO} Algorithm Using the Hypervolume Measure as Selection Criterion},
-  booktitle = {Lecture Notes in Computer Science}
+  booktitle = {Lecture {N}otes in {C}omputer {S}cience}
 }
 
 
 @article{Zitzler2001,
-  title={SPEA2: Improving the strength Pareto evolutionary algorithm},
+  title={SPEA2: Improving the strength {P}areto evolutionary algorithm},
   author={Zitzler, Eckart and Laumanns, Marco and Thiele, Lothar},
   journal={TIK-report},
   volume={103},
@@ -185,7 +185,7 @@
   number = {65},
   pages = {3430},
   author = {Mehmet Hakan Satman and Bahadır Fatih Yıldırım and Ersagun Kuruca},
-  title = {JMcDM: A Julia package for multiple-criteria decision-making tools},
+  title = {JMcDM: A {J}ulia package for multiple-criteria decision-making tools},
   journal = {Journal of Open Source Software}
 }
 
@@ -200,7 +200,7 @@ and S{\'a}nchez-Ram{\'i}rez, Cuauht{\'e}moc
[…]
