+++ paper/paper.bib
@@ -327,7 +327,8 @@ url="https://doi.org/10.1007/978-1-4615-3612-3_4"
 
 @ARTICLE{pymoo,
   author={Blank, Julian and Deb, Kalyanmoy},
-  journal={IEEE Access}, 
+  title={Pymoo: Multi-Objective Optimization in Python},
+  journal={IEEE Access},
   year={2020},
   volume={8},
   number={},
