### 2022-10-26 #4723 @danielskatz
I've suggested minor changes in https://github.com/jmejia8/Metaheuristics.jl/pull/49 and https://github.com/jmejia8/Metaheuristics.jl/pull/50. Please merge these or let me know what you disagree with, then we can continue the process
