### 2022-10-26 #4723 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
