+++ paper/paper.bib
@@ -327,7 +327,7 @@ url="https://doi.org/10.1007/978-1-4615-3612-3_4"
 
 @ARTICLE{pymoo,
   author={Blank, Julian and Deb, Kalyanmoy},
-  title={Pymoo: Multi-Objective Optimization in Python},
+  title={Pymoo: Multi-Objective Optimization in {P}ython},
   journal={IEEE Access},
   year={2020},
   volume={8},
