### 2022-10-26 #4723 @danielskatz
@jmejia8 - sorry, one more thing I thought I fixed but seem to have missed: https://github.com/jmejia8/Metaheuristics.jl/pull/51
